patch hotfix for parens
hotfix 4d84ac73207faada25e9ec3be478d97069fe787c
made it so selecting something and typing a close
paren would resize the paren correctly.

However, this caused typing a close paren without
anything selected to be sized completely wrong
because it's size was being calculated before
it was added to the DOM and hence was being
measured wrong.