add focus/blur to API, don't interfere with DOM focus/blur
before, we'd listen to focus/blur on the MathQuill root span and
trigger it on the textarea, but in some browsers this bubbles, so to
avoid infinite looping, we'd stopPropagation() in the textarea handler.

In ways I don't fully understand, this made detecting focus/blur from
outside MathQuill (via the jQuery focusin/out events) unreliable,
in particular, if you clicked on the math and then left and then clicked
again, there would be a focusin event the first time, but successive
clicks trigger no focusin events even though the textbox is visibly
blurring and focusing.

The solution I settled on is to have an API call to set focus or blur,
and never interfere with actual HTML DOM focus/blur events. Note this
also means the initializing blur() has to be called on the textarea now.

This is what CodeMirror 2 and ACE (Cloud9) both do, as of
https://github.com/marijnh/CodeMirror2/blob/8ba416d5cac2de28a6da9bbc966b8b8761850358/lib/codemirror.js
and
https://github.com/ajaxorg/ace/blob/51b15d35889148f3bd21060bba5cbb32670e36ff/lib/ace/editor.js