give width=height=1px to focusable (tabindex=0) span used by MobiSafari
on Mobile Safari we don't want focusing on a MathQuill editable to pull
up the on-screen keyboard, so instead of a hidden textarea we use a
hidden focusable span, i.e. one with the tabindex attribute.

Now, in most browsers only visible elements, i.e. elements with nonzero
width and height, can be focusable, and in fact in IE calling .focus()
on non-visible element explodes. jQuery <1.7 had a try-catch around
triggering events, and focusing the non-visible span happened to work in
Mobile Safari, but in jQuery 1.7, they changed to checking if the
element had width 0, so it stopped working in Mobile Safari.

This is mostly my fault since non-visible spans aren't supposed to be
focusable in general, so now I'm giving it width=height=1px.