disable textarea in all touch-enabled user agents
This keeps the on-screen keyboard from popping up when MathQuill
focuses. Note that on touch-enabled devices that also have physical
keyboards, this also breaks text entry with the physical keyboard,
because it breaks textarea.val().

Note also that this problem has been present on the iPad before this
commit anyway.

per discussion with @jwmerrill, we're pushing this for now so that
MathQuill-DCG will work on the Nexus 7 (and in theory any other Android
tablet) at least as well as it does on the iPad (i.e. the on-screen
keyboard won't pop up, but external keyboards won't work).