fix #135 in Safari 5.1 can't type over selection
In safari, there's a brief moment when typed text is marked as selected, as mentioned by @laughinghan in
https://github.com/desmosinc/mathquill/commit/e9fa48a2d708ab567f5205c181583784d548d243

It turns out that waiting one cycle is enough to deselect the text. The "flush" bound to "input" was firing just after keypress, but in the same cycle, overriding the defer triggered in onKeyPress. Removing "input" from the list of actions that triggers a flush fixes this problem.

I think this is safe, because at worst it delays whatever action was attached to something triggered by "input" by one cycle.