reverting the input removal
It turns out that *not* calling this popText immediately breaks all sorts of things in DCG, most notably the alignment info that mathquill reports to the outside world. As a result, typing aligns things incorrectly both horizontally and vertically if a flush isn't called.

This patch catches the weird safari bug by trying inline (before the flush), and then trying once more after the flush (by that built-in timeout). Fixes the safari bug (#1266 in knox), but doesn't get around the aforementioned bug (horizontal/vertical alignment) in the circumstance that this catch fires (Safari after select, #1266).

Uggh.