chack charCode instead of using regex in makeCmd #optimize
interestingly, Chrome 27.0.1453.93 (Official Build 200836) normally
optimizes regexes after the first use; but in this case, it's optimized
for successive calls to makeCmd but not the first call to makeCmd *each
renderSliderLatex call*, perhaps because it's in an inner function
that's recreated every time? In any case, it adds ~0.1ms to each call to
renderSliderLatex, almost 10% of the running time now.