MathElement::seek now minimizes Euclidean distance
Replaces Math{Block,Command}::seek, but not Symbol::seek.

The algorithm is reasonably simple (start with guesses of a nearby
Point, a containing box, and nearby boxes, and until there's a Point
closer than the closest coords in any box, look in the closest box for
more guesses, as suggested by @ehberger), and is expressed reasonably
compactly at the bottom of MathElement::seek, but add.node and
MathBlock::seekPoint are both much more verbose than I'd like.

In order to estimate the Euclidean distance from the sought coordinates
to each Point with just DOM measurement (rather than inserting the
cursor at that Point, which is a massively slower DOM manipulation),
added 2 methods:
- MathBlock::expectedCursorYInside
- MathCommand::expectedCursorYNextTo
Many commands override expectedCursorYNextTo, depending on their
vertical-align.