Vertical squeezing/stretching of cursor handle works!
Made cursor handle into an element (you can't select pseudo-elements
with JS to style dynamically, for example, a transform that has to be
computed, separately from the cursor itself), added to `Cursor` as the
`.handle` property when there's a handle at all (after a touch event),
0 when no handle (by default and after a mouse event). Default to 0 and
not null so `cursor.handle[0]` doesn't throw.

translateX(-.5px) (which was there because want the cursor handle's point
to be centered in the 1px-wide cursor, in case that wasn't clear) moved
to the .handle element because when on the pseudo-elements,
.css('WebkitTransform', '') doesn't work (no idea why).

I also tried `weight = 1/Math.sqrt(dist)`, but natural log falls of
slower at first and faster later; with inverse sqrt, the handle would
still move around substantially when your finger was way across the
screen.

I couldn't get the math of skewX working, and getting just scaleY
working was trickier than I expected, so I'm committing now that I got
just scaleY to work.