Fix annoying occasional white line within handle
...between the triangular tip and rectangular body. The line, when it
shows up, is always much less than a screen pixel tall (even on the
Retina iPad where there are clearly multiple screen pixels per CSS
pixel), I'm guessing some kind of floating-point rounding error in the
hardware-accelerated graphics.

It was a problem before the stretching/squeezing, like it would show up
when the cursor was in full-size text but not in an exponent, for
example, but it's much more annoying when it flashes because the handle
is being squeezed/stretched continuously as the finger moves.