Faster frontier with memory-efficient "priority queue"
I was surprised to see that in Chrome, at least, sorting the frontier
was taking dozens of microseconds (several microseconds, sure, but
dozens? Tens of thousands of nanoseconds to sort a definitely
size < 100 array?). As a knee-jerk reaction, I tried using an array-
backed binary heap implemented in ~100 lines of JS, but that was slower,
which in retrospect I should've expected.

With a little thought I realized I needed to make things simpler, not
more complicated, so now the frontier is a plain array, finding the
closest point or box is a linear-time (but low-overhead, cache-friendly)
traversal over the array.