expectedCursorY{NextTo,Inside} uses cached measurements
In cases like SupSub and Fraction, where the relevant DOM measurement is
of a DOM node without a corresponding MathQuill edit tree node, pass in
a mock edit tree node with the slots the cache cares about. Hacky, but
it goes away soon.

The expectedCursorY{NextTo,Inside} calls aren't that slow, this commit
is prep for the next commit to short-circuit jQuery and call
getBoundingClientRect directly, which should be both faster and happen
less frequently since it gives us *exactly* the information we actually
care about. The dumb thing is, getBoundingClientRect, like .client{X,Y}
on mouse and touch events, returns coords relative to the viewport,
whereas .page{X,Y} and .offset().{left,top} are relative to the
document. This commits makes all DOM measurement in seek go through
getBox, which can then be converted to just getBoundingClientRect
instead of multiple jQuery calls.