pageX/Y => clientX/Y, jQuery::offset() => getBoundingClientRect
Short-circuit jQuery::offset() and directly call getBoundingClientRect,
because jQuery adds significant overhead, and .offset() throws away
information we want and end up having to make more DOM measurements to
get. However, they're not equivalent: jQuery::offset() returns
coordinates relative to the document, getBoundingClientRect() returns
coordinates relative to the viewport, which is dumb because that
coordinate system changes if you scroll. Assuming no scrolling happens
during a mouse-drag-to-select or touch-drag-to-place-cursor, though, we
can then just use the viewport-relative .clientX/Y properties of the
{mouse,touch}move event object, instead of the document-relative
.pageX/Y.

I changed a lot of variable names to make clear we're using
viewport-relative 'client' coordinates consistently, so much of this
diff is easier to read with --word-diff.

It's definitely nice that BoundingClientRects have pretty much exactly
the information we actually want, there are so many instances of
simplifying `.y() + .height()` to `.bottom` or `.x() + .width()` to
`.right`; in other places, `.y() + .height()/2` is replaced by the no
more complicated `(.top + .bottom)/2`.

Noticeable performance improvement in visual.html on the iPad! #optimize