Don't look "through" cursor handle, DOM manip slow
Whoops, that's why seeking with cursor handle on iPad was so, so much
slower than seeking with a mouseclick, I had only been testing on the
iPad if it felt faster but trusting the performance logs in Chrome to
log the same performance bottlenecks.

This now means it's possible [1] for the coordinates passed to .seek()
to be outside the bounding rectangle of the node that Cursor::seek()
chooses to call .seek() on, so MathElement::seek() now doesn't assume
that containers necessarily contain the coordinates, but it's more
efficient if they do.

[1]: In generality you can see this is possible because the cursor
handle violates the assumption that an event's coordinates must be
within the bounding rectangle of the event's target. In particular,
this is possible in a "race condition"-like edge case where if your
finger was dragging on the actual math, so moving the cursor directly
and not dragging the handle, and you move your finger fast enough to
cross the distance to the handle during the delay between events, so a
touchmove event fires with your finger directly over the cursor handle,
then elementFromPoint returns the cursor span and Cursor::seek would
call .seek() on the containing MathBlock (because the cursor span's
parent DOM node would be a MathBlock span), which in that case might
not actually contain the sought coords within its bounding rectangle.
This is a transient edge case, though, and seek wouldn't misbehave
too badly. The other edge case where this is possible is if the cursor
is in the lowest position possible; then moving your finger further down
is plenty likely, not at all transiently, to pass over the cursor handle.