Fix touchstart from outside MathQuill
Whereas mousemove/up events fire on and bubble up from the element the
mouse is over at that moment, touchmove/end events fire on and bubble up
from the original element that touchstart was fired on [1]. Hence,
whereas on mousedown we attach mousemove/up event handlers to the
document where mousemove/up events from any element should bubble up to,
on touchstart we attach touchmove/end event handlers to the original
element the touchstart was fired on, which we previously assumed to be a
descendant of `this` (the element this handler is being attached to),
a bad assumption if the touchstart is being passed from outside.

Also update [test case: trigger mousedown from elsewhere][3ee6a34] to
test touchstart.

[3ee6a34]: 3ee6a34ae560d2d23a2cd3b5170b9a8d340c913a

[1]:
"[I]n a normal `mousemove` event, the node passed in the `target`
attribute is usually what the mouse is currently over. But in all iPhone
touch events, the target is a reference to the originating node."
  - probably the most widely cited touch events resource, before being
    updated:
    http://www.sitepen.com/blog/2008/07/10/touching-and-gesturing-on-the-iphone/
"The event's target is the same element that received the touchstart
event corresponding to the touch point, even if the touch point has
moved outside that element."
  - MDN:
    https://developer.mozilla.org/en-US/docs/Web/API/TouchEvent#touchmove
"The target of this event must be the same Element on which the touch
point started when it was first placed on the surface, even if the touch
point has since moved outside the interactive area of the target
element."
  - W3C spec, as of 09 May 2013:
    http://www.w3.org/TR/2013/PR-touch-events-20130509/