Don't block scroll when touch swipe starts in MathQuill
Argghhhhh. Want touch swipes to scroll the page, even if they start in
MathQuill (so that it's possible to scroll an expression list filled to
the brim with nothing but MathQuills) but touch taps to place the
cursor. Now, with mouse events, if all you care about is a logical
"click" on the current platform, you just listen for the click event and
the browser will take care of whether a particular sequence of
mousedown-mousemoves-mouseup is a click or not. With touch events,
there's no equivalent logical "tap" event, and while legacy mouse events
do fire if you don't preventDefault() on touchstart, Safari on iOS waits
a glacial ~350ms after a tap in case you're actually doing the
double-tap gesture (which doesn't fire any events). As a result there
are dozens of blog posts about and shitty JS libraries for "fast clicks"
across the Web, including [this widely cited Google Developers
article][1], that vary widely in how much your finger is allowed to
wiggle during the tap.

Desmos already has a tap-event library, so for consistency I just expose
an API call for the client to tell MathQuill when it thinks a logical
tap has happened. Given the touchstart target and the touchend
coordinates, MathQuill takes care of ignoring the tap if it was on the
cursor handle and doing document.elementFromPoint() on the touchend
coordinates and dealing with whether it's inside the math field or not.
(This required moving elAtPt() to the MathQuill-global scope and adding
a `root` block argument.)

But wait, there's more! Not only are the legacy mouse events unhelpful,
they're misleading because there is no way whatsoever to tell them apart
from real mouse events and no way to get them not to fire except calling
preventDefault(), which breaks touch-swipe-to-scroll. So with
heart-wrenching dispair, I expose an API call to ignore the next mouse
events (and hence not hide the cursor handle) up to a timeout, meant to
be called on touchend before every expected fake mousedown, regardless
of whether there was a logical "tap".

The test case has a trivial notion of logical "tap", which is if no
touchmoves happen between the touchstart and touchend. Note that it
calls .mathquill('ignoreNextMousedown', 1000) after every touchend.
Note also that the test client now needs to take care not trigger
duplicate mousedowns on MathQuill, as MathQuill will only ignore the
first.

The only non-JS change to the test case was to hide the gray tap
highlight box.

[1]: https://developers.google.com/mobile/articles/fast_buttons