Embiggen cursor handle hit box / touch target
Pseudo-element structure of cursor handle completely changed,
from:
  - the handle span being a 0x0 box at the bottom of the cursor line
  - the ::before pseudo-element being the triangle tip
  - the ::after pseudo-element being the square body
to:
  - the handle span being the square body
  - the ::before pseudo-element being the triangle tip, still
  - the ::after pseudo-element being the invisible embiggened hit box /
    touch target

The reason for the new structure rather than just having the handle span
be the hit box is that the pseudo-elements' coordinates are relative to
the handle span, and the hit box's size and coordinates needed to be
(and likely will need to be) tweaked repeatedly, so by making the
::after pseudo-element be the hit box, its size and coordinates can be
tweaked without affecting the rest of the handle.

Note that though the ::before pseudo-element is still the triangle tip,
it's coordinates have changed since its coordinates' origin, the handle
span, moved.

The transform origin y-coordinate should technically be -20px above the
top of the handle span, but 5px fudge factor makes it feel better.

The subpixel gap fudge factor (which is now commented) was moved from
the handle body to the triangle tip, because the tip (::before pseudo-
element)'s coordinates are relative to the body (handle span).