Fix mouse-selecting erroring sometimes
Fix mouse-selecting between two places whose lowest common ancestor is
the root block when one of the two places is not the root block.

Bug intro'd in 1e2965072a82ba9967bbb2175dc0772dba25409f, which hastily
attempted to patch an infinite loop.

This needs test cases but needs to be factored out to not need extensive
mocks to test, but it was factored out and tested and the infinite loop
fixed in mainline MathQuill in f70ecb91c608f933e548fd189d3a405891750fdd,
so I'm just gonna leave this be.