First pass horizontal panning instead of wrapping
Like what <input/>'s do, since wrapping behavior for math is
ill-specified.

So for dumb reasons, CSS2.1 specifies that the baseline of an
inline-block that is overflow:hidden is its bottom margin edge [1]
(not for no reason, I admit [2]). For now just set vertical-align:middle
to look reasonable for the common case, eventually I'll adjust the
vertical-align numerically in JS (after all, I already have
Node::expectedCursorYInside).

[1]: http://lists.w3.org/Archives/Public/www-style/2013Jul/0188.html
[2]: http://lists.w3.org/Archives/Public/www-style/2013Jul/0471.html

Need extra 1px at right edge lest the cursor, which has
margin-right:-1px, overflow and be hidden.

Shrink textarea to smallest nonzero width and height so it won't affect
the minimum scroll{Width,Height} of the container. This took quite a
while to track down the other day, the only symptom was that the
scrollWidth was a weirdly big number, there's not a lot of people
talking about what affects scrollWidth. (Must be nonzero so that
textarea is visible and hence focusable.)

Known issues:
- ignores mouse
- keyboard selection errors
- in spite of adding .scrollHoriz() call inside .onText(), totally
  doesn't work when typing
    (conjecture: typing is re-focusing textarea which is at the left
     edge of container scroll area?)