Horizontal panning uses just cursorX
cursorRect.left should be the same as cursorRect.right anyway, and in
fact that's the only reason this worked; with selection, don't want to
show both sides of the selection because can't, necessarily, instead
want to show whichever side of the selection the cursor is at.

Recommend --word-diff-regex='\w+|.' to view diff.