Improve horizontal panning behavior #fuckyeah
First pass implementation of the behavior I want; the implementation
could almost certainly be more elegant and de-duplicated, but for how
the intuitive the behavior is it was surprisingly tricky to put into
code.

I'm actually really happy with the reward for the attention to detail
here. The basic idea, I guess, is to scroll by a minimum amount while
maximizing the amount of the selection shown, and while keeping the
cursor within the inner bounds of the textbox (i.e. 20px from the
edges).