Move overflow:hidden to mathquill-root-block
Needed to fix cursor touch handle, which needs to be inserted into DOM
outside overflow:hidden container.

Happens to also fix flicker when typing past inner cursor bound, since
textarea that keeps being re-scrolled-to is outside
mathquill-root-block.

Needed to change margin-right to padding-right to avoid 1px shift when
cursor is at the right edge of a textbox that overflows, not entirely
sure why but conjecture is margin-collapse.