Fix 1px shift when cursor at right edge in Firefox
padding-right intro'd in 0ed4a3cde110918cd92648e76de3a21955fe7fd2 to fix
1px shift when cursor goes to the end of a textbox that overflows;
the conjecture in that commit message is wrong, and the 1px shift is
because the cursor has a negative `margin-right` so when it's at the
right edge, it overflows which increases the `scrollWidth` and since
`.scrollHoriz()` is pushing to scroll as far right as possible, the
entire editable scrolls to the right by 1px. The padding-right actually
only appeared to fix the problem due to a WebKit bug, so we need a real
way to add 1px to the scrolling area.

Firefox Bugzilla thread, marked invalid, claiming the behavior was
intro'd in Firefox 12 deliberately, matches IE10, and was matched by
Opera in Feb. 2013 whence WebKit is the odd one out:
https://bugzilla.mozilla.org/show_bug.cgi?id=748518

https://github.com/desmosinc/mathquill/commit/0ed4a3cde110918cd92648e76de3a21955fe7fd2#commitcomment-4493437