Move cursor handle outside overflow:hidden box
Cursor handle is typically completely outside the main box and is
completely cropped out, so at first I thought I had somehow totally
broken the cursor handle.

Apparently there's no way for a descendant of an overflow:hidden element
to be "above" the clipping area, nor any CSS-only way to "lock" the
position of an element relative to another except its offset parent,
so have to use JS to re-position the cursor handle whenever the cursor
moves.

getBoundingClientRect's results are after applying transforms (which
surprised me because afaik that makes it the only transform-aware
CSSOM API ever), but that's not necessarily consistent cross-browser[1],
and I didn't feel like figuring out the calculation to reverse the
transform, and apparently there's no API to get pre-transform coords, so
I chose to have an extra 'anchor' span that I would measure and position
so its bottom lines up with the bottom of the cursor, that would contain
and be the offset parent of the handle span which could then be
transformed willy-nilly. As a bonus, the relative position of the handle
and the cursor remains entirely in CSS.

[1]: this apparently this used to vary by browser, with WebKit returning
post-transform values and Gecko returning pre-transform values[2], but
Gecko eventually matched WebKit[3], though www-style never appeared to
come to a consensus over many scattered discussions[4].

[2]: http://bugs.jquery.com/ticket/8362
[3]: https://bugzilla.mozilla.org/show_bug.cgi?id=591718
[4]: http://lists.w3.org/Archives/Public/www-style/2011May/0385.html

(The other option I could see is making the cursor handle position:
absolute rather than relative, so I wouldn't need to measure its coords,
just its offset parent's. But then I'd either have to set the
`.mathquill-editable` to position:relative so it'll be the offset
parent, potentially conflicting with a style set by the host page; or
insert the handle inside the `.textarea` span, but that's a conflation
of purpose and while its purpose could be expanded, I was meaning to get
rid of it anyway, turns out setting position:absolute on the textarea
and leaving top and left as auto does exactly what I want.)

As an optimization, store current top and left rather than make DOM
measurements every time.

Handle no longer hides when cursor hides by virtue of being inside it,
so have to manually hide handle when cursor hides. Note that before, the
handle would show again when the cursor was shown again unless the
handle was hidden. Now it won't show again until the .showHandle() is
called, but it doesn't matter because whenever the cursor is shown
again, it hides the handle (the only time the Cursor::show() doesn't
hide the handle is on touchend, which is just getting the cursor to
resume blinking).

It is nice that the handle is now permanent, so the touchstart event
handle can be attached to it permanently.

Fixes issue brought up in this comment:
https://github.com/desmosinc/mathquill/commit/dd83c9f73315cfc28db6bd703ff922515247050b#diff-b2d5fff9867d8d6456e172d381a87d49R4