Horiz pan works for clicking/selecting/handle-dragging
Fully functional for clicking or tapping anywhere, for mouse-selecting,
and for dragging the cursor handle. Performance in Safari on iOS when
dragging the handle feels shoddy but isn't unusable.

Note that we want to cache the clientRect's for the entire duration of a
mouse selection or handle drag, across many mouse/touchmove's, but we
only want to cache the scrollLeft for a single event thread. Luckily we
know that all the calls to cachedClientRect are by seek, so cache it
there and invalidate it after and explicitly mark the implicit
dependency on scrollLeft being cached.

Also note that we deliberately call root.scrollHoriz() before
cursor.selectFrom() when mouse-selecting, so we'll hit the no-selection
case in scrollHoriz and scroll to the seek'd cursor position rather than
end of the selection, which reduces the scrolling speed because the
cursor often goes into e.g. the near end of a wide fraction, whence the
end of the selection would be at the far end of the wide fraction.

It's kinda dumb that ClientRect's properties are all read-only.

I factored out the minimalist touchtap implementation so I could test
the cursor handle in the overflowing box, too.