Fix cursor handle becoming "detached" from cursor
...if you drag the handle past the inner cursor boundary and then let
go; the cursor seeks past the inner cursor boundary, the handle is
repositioned under it, and then the scroll animates, shifting the
cursor. When this is switched to CSS animations, can just listen to an
'animationend' event; for now, just set a timeout for after the
longest the animation could take (could be earlier than this if most
recent touchmove was long before this touchend).