Fix cursor handle when animating scroll
While the finger is touching the handle, on animation end re-stretch the
handle; on touch end, on animation end reposition/"snap back" handle to
under cursor.

Recommend -w to view diff because of factoring out visualHapticFeedback
function.

Store .onAnimationEnd on root, not on cursor handle, because after all
that's what's animating. And fix how 'complete' handler is passed to
.animate(); 639f291a54221e68ddd2185a276d901288baf109 never worked.