\editable works; mouseEvents uses closest root block
Completely replace the old RootMathCommand-descended, jQadd-overriding
\editable. Instead, LatexCmds.editable::finalizeTree initializes a
RootMathBlock as the sole child of the \editable MathCommand to match
how createRoot initializes normal root blocks: the containing span's
mathquill-block-id attribute is the .id of the RootMathBlock, whose
.jQ is the .mathquill-root-block inside the containing span (it's weird
but is like that so publicapi.js works).

The main reason for the big createRoot() refactor in the previous commit
is that the mousedown handler attached to both the root span of the
static math and the root of the inner editable stepped on each other's
toes. I originally did stopPropagation, but that broke DCG
(desmosinc/knox#900) and is bad anyway, the right thing is for mousedown
to be a delegated event, with the handler only attached to the actual
root of the MathQuill tree (the root of the static math), which then
uses the closest ancestor that's a root math block to determine
editability, to attach the mousemove handler, focuses its textarea, and
seeks its cursor.

Note that it's important to attach the mousemove event handler to the
closest ancestor root block, lest the target node not descend from the
cursor's root, breaking Cursor::seek. (The document-root mousemove
handler ignores the target node.)