Fix cursor handle under later positioned elements
Want only positive z-index elements to appear above cursor handle.

25bc721519010a4dcf2e6b873fb5961f724e6127 introduced a bug that is
difficult to see from its diff: [`.mq-cursor` has `z-index:1`][1], which
when it contained the handle meant that the handle would appear in front
of even positioned elements that appear later in the same stacking
context (I had thought that was because the handle was
`position:absolute`, but nope, [positioned elements are all in the same
layer, painted back-to-front in order][2]).

The result was the handle would appear in front of subsequent
`position:static` elements but not positioned elements, and hence
appeared fine in `test/visual.html` but was largely broken in DCG.

[1]: https://github.com/desmosinc/mathquill/commit/25bc721519010a4dcf2e6b873fb5961f724e6127#diff-b2d5fff9867d8d6456e172d381a87d49L19
[2]: http://reference.sitepoint.com/css/stacking