fix chrome/window bug with -webkit-text-stroke
On Chrome on Windows, when the font shrunk +'s, -'s, ='s were losing some horizontal lines. This seems to be a known bug in the Chrome rendering engine, and is affecting all Google Web Fonts also (they default to woff):
https://code.google.com/p/chromium/issues/detail?id=137692

Of the many recommended hacks / fixes, the one that looks best is to add the -webkit-text-stroke css rule. Recommended size is .35px. I found that .3px looks best.

We'll want to revert this commit when Chrome finally gets it's freaking act together.